[order] request order improvements
